backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
yarl==1.9.4
